+++ src/framework/listener/__init__.lua
@@ -6,6 +6,7 @@
 
     Member:
       _object_ - object using the listener
+      _registry_ - record what event is registered.
 
     Function:
       new() - create a new instance
@@ -26,13 +27,23 @@
         ... - args you want to put.
 --]]
 
-local require = require
-local cls = require 'std.class'("Listener", require 'framework.listener.event')
-local subject = pcall(require, 'framework.listener.subject')
+local require, xpcall, select = require, xpcall, select
+local cls = require 'std.class'("Listener")
+local error = function() return false end
+local subject = select(2, xpcall(require, error, 'framework.listener.subject'))
+local template = select(2, xpcall(require, error, "framework.listener.template"))
 
 function cls:_new(obj)
-    local this = self:super():new()
-    this._object_ = obj
+    local this = {
+        _event_ = {},
+        _object_ = obj,
+        _registry_ = {},  -- 記錄有沒有重複對主題註冊事件
+    }
+
+    if subject then
+        subject.addListener(this)
+    end
+
     return this
 end
 
@@ -40,19 +51,77 @@ function cls:object()
     return self._object_
 end
 
-function cls:addBroadcast(name, arg, callback)
-    if subject then
-        subject:addEvent(name, self._object_, arg, callback)
+function cls:isRegistered(name)
+    return self._registry_[name]
+end
+
+function cls:add(name, arg, callback)
+    local type = type
+
+    if not(type(name) == 'string' or type(name) == 'table') then
+        return self
+    end
+
+    local event
+    -- NOTE: 解析只有名字或是一個事件的情況
+    if not (arg or callback) then
+        if type(name) == 'table' then
+            event = name
+            name = name.name
+        elseif template and template[name] then
+            event = template[name]
+        else
+            return self
+        end
+    else
+        event = {name=name, arg=arg, callback=callback}
+    end
+
+    if not self._event_[name] then
+        self._event_[name] = {}
+    end
+
+    self._event_[name][#self._event_[name] + 1] = event
+
+    if not self._registry_[name] and subject and subject.add(self, name, arg) then
+        self._registry_[name] = true
     end
 
     return self
 end
 
 function cls:onTick(name, ...)
+    if not self._event_[name] then
+        return false
+    end
+
     -- 把 self 加入到參數列裡
     local arg = {...}
     arg[#arg+1] = self
-    return self:super().onTick(self, name, table.unpack(arg))
+
+    local unpack = table.unpack
+    for _, event in ipairs(self._event_[name]) do
+        event.callback(unpack(arg))
+    end
+
+    return self
+end
+
+function cls:iterator(name)
+    if not self._event_[name] then
+        return
+    end
+
+    local i = 1
+    return function()
+        if i > #self._event_[name] then
+            return
+        end
+
+        local arg, callback = self._event_[name][i].arg or "", self._event_[name][i].callback
+        i = i + 1
+        return arg, callback
+    end
 end
 
 return cls