+++ src/framework/listener/template/__init__.lua
@@ -1,0 +1,3 @@
+return require 'framework.load'("framework.listener.template", {
+    "test"
+})