+++ src/framework/listener/template/test.lua
@@ -1,0 +1,6 @@
+return {
+    name="測試",
+    callback=function()
+        print("testing...")
+    end
+}