+++ src/test/framework/listener.lua
@@ -1,20 +1,26 @@
 local a = require 'framework.listener':new(1048591)  -- 遊戲裡的血法師的id
 print(a:object())
-a:addEvent("測試", function() print(2) end):addEvent("測試", function(i) print(i) end):addEvent("測試", function(i, this) print(i + this:object() * 2) end)
+a:add("測試", nil, function() print(2) end):add("測試", nil, function(i) print(i) end):add("測試", nil, function(i, this) print(i + this:object() * 2) end):add("測試")
 a:onTick("測試", 10)
-a:addBroadcast(
+
+for arg, callback in a:iterator("測試") do
+    print(arg, callback)
+end
+
+-- war3 測試
+a:add(
     "單位-受到傷害",
     "GetTriggerUnit",
     function(u)
         print(u .. " is damaged. Please help him.")
     end
-):addEvent("單位-受到傷害", function(u) print("I'm good") end)
+):add("單位-受到傷害", nil, function(u) print("I'm good") end)
 
 local b = require 'framework.listener':new(1048590) -- 遊戲裡的強盜的id
-b:addBroadcast(
+b:add(
     "單位-受到傷害",
     "GetEventDamageSource",
     function(u)
         print(u .. " hurt someone.")
     end
-):addEvent("單位-受到傷害", function(u) print("I'm bad") end)
+):add("單位-受到傷害", nil, function(u) print("I'm bad") end)